requirements/dev.txt
appdirs==1.4.4            # via virtualenv
asgiref==3.2.10           # via django
astor==0.8.1              # via wemake-python-styleguide
astroid==2.4.2            # via pylint
attrs==19.3.0             # via flake8-bugbear, flake8-eradicate, pytest, wemake-python-styleguide
bandit==1.6.2             # via flake8-bandit
cfgv==3.2.0               # via pre-commit
click==7.1.2              # via pip-tools
coverage==5.2.1           # via django-coverage-plugin, pytest-cov
darglint==1.5.2           # via wemake-python-styleguide
distlib==0.3.1            # via virtualenv
dj-database-url==0.5.0    # via environs
dj-email-url==1.0.1       # via environs
django-cache-url==3.1.2   # via environs
django-coverage-plugin==1.8.0  # via -r requirements/dev.in
django-extensions==3.0.5  # via -r requirements/dev.in
django==3.1               # via -r requirements/./base.in, djangorestframework, model-bakery
djangorestframework==3.11.1  # via -r requirements/./base.in
docutils==0.16            # via restructuredtext-lint
environs[django]==8.0.0   # via -r requirements/./base.in
eradicate==1.0            # via flake8-eradicate
filelock==3.0.12          # via virtualenv
flake8-bandit==2.1.2      # via wemake-python-styleguide
flake8-broken-line==0.2.1  # via wemake-python-styleguide
flake8-bugbear==19.8.0    # via wemake-python-styleguide
flake8-commas==2.0.0      # via wemake-python-styleguide
flake8-comprehensions==3.2.3  # via wemake-python-styleguide
flake8-debugger==3.2.1    # via wemake-python-styleguide
flake8-django==1.1.1      # via -r requirements/dev.in
flake8-docstrings==1.5.0  # via wemake-python-styleguide
flake8-eradicate==0.3.0   # via wemake-python-styleguide
flake8-isort==3.0.1       # via wemake-python-styleguide
flake8-polyfill==1.0.2    # via flake8-bandit, pep8-naming
flake8-quotes==2.1.2      # via wemake-python-styleguide
flake8-rst-docstrings==0.0.12  # via wemake-python-styleguide
flake8-string-format==0.2.3  # via wemake-python-styleguide
flake8==3.8.3             # via flake8-bandit, flake8-broken-line, flake8-bugbear, flake8-commas, flake8-comprehensions, flake8-debugger, flake8-django, flake8-docstrings, flake8-eradicate, flake8-isort, flake8-polyfill, flake8-quotes, flake8-rst-docstrings, flake8-string-format, wemake-python-styleguide, yesqa
gitdb==4.0.5              # via gitpython
gitpython==3.1.7          # via bandit
identify==1.4.25          # via pre-commit
iniconfig==1.0.1          # via pytest
isort[pyproject]==4.3.21  # via flake8-isort, pylint
lazy-object-proxy==1.4.3  # via astroid
marshmallow==3.7.1        # via environs
mccabe==0.6.1             # via flake8, pylint
model-bakery==1.1.1       # via -r requirements/dev.in
more-itertools==8.4.0     # via pytest
mypy-extensions==0.4.3    # via mypy
mypy==0.782               # via -r requirements/dev.in
nodeenv==1.4.0            # via pre-commit
packaging==20.4           # via pytest
pathspec==0.8.0           # via yamllint
pbr==5.4.5                # via stevedore
pep8-naming==0.9.1        # via wemake-python-styleguide
pillow==7.2.0             # via -r requirements/./base.in
pip-tools==5.3.1          # via -r requirements/dev.in
pluggy==0.13.1            # via pytest
pre-commit==2.6.0         # via -r requirements/dev.in
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via flake8, flake8-bandit, flake8-debugger
pydocstyle==5.0.2         # via flake8-docstrings
pyflakes==2.2.0           # via flake8
pygments==2.6.1           # via -r requirements/dev.in, wemake-python-styleguide
pyhamcrest==2.0.2         # via -r requirements/dev.in
pylint-django==2.3.0      # via -r requirements/dev.in
pylint-plugin-utils==0.6  # via pylint-django
pylint==2.5.3             # via pylint-django, pylint-plugin-utils
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.0        # via -r requirements/dev.in
pytest-django==3.9.0      # via -r requirements/dev.in
pytest==6.0.1             # via -r requirements/dev.in, pytest-cov, pytest-django
python-dotenv==0.14.0     # via environs
pytz==2020.1              # via django
pyyaml==5.3.1             # via bandit, pre-commit, yamllint
restructuredtext-lint==1.3.1  # via -r requirements/dev.in, flake8-rst-docstrings
six==1.15.0               # via astroid, bandit, django-coverage-plugin, packaging, pip-tools, virtualenv
smmap==3.0.4              # via gitdb
snowballstemmer==2.0.0    # via pydocstyle
sqlparse==0.3.1           # via django
stevedore==3.2.0          # via bandit
testfixtures==6.14.1      # via flake8-isort
tokenize-rt==4.0.0        # via yesqa
toml==0.10.1              # via isort, pre-commit, pylint, pytest
typed-ast==1.4.1          # via mypy
typing-extensions==3.7.4.2  # via mypy, wemake-python-styleguide
virtualenv==20.0.30       # via pre-commit
wemake-python-styleguide==0.14.1  # via -r requirements/dev.in
werkzeug==1.0.1           # via -r requirements/dev.in
wrapt==1.12.1             # via astroid
yamllint==1.24.2          # via -r requirements/dev.in
yesqa==1.2.1              # via -r requirements/dev.in

# The following packages are considered to be unsafe in a requirements file:
# pip
